pyproject.toml
[project]
name = "radipop_utils"
version = "0.1"
requires-python = ">=3.10"
description = "The radipop_utils package contains useful functions for extracting most important radiomics features from CT images used for HVPG prediction as well as time-to-event prediction (outcome)."
readme = "README.md"
authors = [
    { name = "Martin Luther Watzenböck", email = "martin.watzenboeck@meduniwien.ac.at"},
    { name = "Department of Radiology; Medical University of Vienna" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Radiomics",
]
keywords = [
    'medical image analysis',
    'radiomics',
]
dependencies = [
    "python-dotenv",
    "numpy>=1.21.0,<2.0.0",
    "pandas",
    "seaborn",
    "matplotlib",
    "pydicom>=2.3.0,<3.0.0",
    "nibabel",
    "dcm2niix",
    "SimpleITK>=2.2.1",
    "pillow",
    "pyradiomics",
    "numba",
    "scipy",
    "scikit-learn==1.3.0",
    "scikit-optimize>=0.9.0", 
    "scikit-image>=0.19.3",
    "tqdm",
    # "TotalSegmentator",   # This is not neccesary for the main code part. Can be installed optionally later
    #"torch>=2.1.2",    
    "ipykernel",
    "jupyter",
    "openpyxl",
    "pyyaml",
    "dicom2nifti<2.6.0"
]

[project.scripts]
radipop_extract_features = "radipop_utils.tools.extract_features_entry_point:main_function"
radipop_extract_features_many_times = "radipop_utils.tools.extract_features_from_many_patients_entry_point:main_function"
radipop_suggest_binwidth = "radipop_utils.tools.suggest_radiomics_binwidth_entry_point:main_function"
radipop_training_and_hyperparams_search = "radipop_utils.tools.training_and_hyperparams_search_entry_point:main_function"
radipop_train_with_optimal_hp = "radipop_utils.tools.train_with_optimal_hp:main_function"
radipop_evaluate_model = "radipop_utils.tools.evaluate_model_entry_point:main_function"
radipop_segment_and_predict = "radipop_utils.tools.auto_segmentation_and_prediction_entry_point:main_function"
radipop_combine_radiomics_with_scalar_target_and_split = "radipop_utils.tools.combine_radiomics_with_scalar_target_and_split:main_function"
radipop_pipeline_dcm_separation_dcm2nii_totalseg = "radipop_utils.tools.pipeline_dcm_separation_dcm2nii_totalseg_entry_point:main_function"
radipop_combine_masks_and_extract_features = "radipop_utils.tools.combine_masks_and_extract_features_entry_point:main_function"

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "pre-commit"
]
torch = [
    "torch>=2.1.2",
    "TotalSegmentator"
]
dicom2nifti = [
    "dicom2nifti<2.6.0"
]
all = [
    "torch>=2.1.2",
    "TotalSegmentator",
    "dicom2nifti<2.6.0"
]

[build-system]
requires = ["setuptools>=67.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.codespell]
skip = '.git,*.pdf,*.svg'
#
# ignore-words-list = ''


[tool.setuptools.packages.find]
where = ["."]


[project.urls]
homepage = "https://github.com/mlwatzenboeck/RADIPOP"

